[jml4] Further support for assignable clauses.        This is the first of "two or three parts"        Patch submitted by Jooyong Lee.
<Overview>

The objective of this patch is to add extra data of user-given
assignable clauses that is likely to be useful for later steps of
compilation and verification. More specifically, JDT's Reference
objects corresponding to assignable locations will become available by
this patch. This makes it possible to know assignable location's type,
binding information such as its modifier and so on, location's
receiver, and so forth.

We are going to split this assignable-clause patch into two or three
parts. This part of the patch focuses on dealing with store refs
except for store-ref keywords (i.e., \nothing, \everything and
\not_specified) and informal description. These will be handled in the
next patch. Currently, we can handle the following store-refs:

  1. single expressions: e.g., x
  2. qualified expressions: e.g., x.y.z
  3. array expressions: e.g., a[1], a[0..2], a[*]
  4. "this" and "super" expressions: e.g., this.i, super.i, A.this.i 
  5. combinations of the above: e.g., x.a[*], x.a[0..2].y, this.a[1].y
  6. wild-card expressions where wild cards are appended to the end of
     the above expressions: e.g., x.*, x.y.*, a[0..2].*, this.a[1].*
     
Next parts of the patch will handle:

  1. store-ref keywords
  2. data groups
  3. error reporting
  4. informal description

====================================
<Changes>

[JmlStoreRefExpression.java]

1. This is where most additions are made.

2. Field assignableReferences is added to hold an array of Reference
objects matching this store-ref expression. Note that there can be
more than one Reference object due to a wild card and data groups.

3. Type and binding resolution takes place at resolve method. This
method calls either resolveNameRef or resolveFieldRef method where
most of the work is performed.

4. Resolution is performed by creating appropriate Reference objects
and calling their resolveType methods. For example, expression x.y is
resolved by creating an instance of QualifiedNameReference and calling
its resolveType method. Similarly, x.y[1].z is resolved by creating
in a row an instance of QualifiedNameReference for x.y and
ArrayReference's instance for x.y[1] and FieldReference's instance for
x.y[1].z before invoking FieldReference's resolveType method.

5. This resolution procedure involves lexing and parsing an expression
given with an assignable clause. It may look unusual at first to lex
and parse an expression again. This was the choice, however, to extend
the original assignable-clause implementation that generates at JDT's
parsing stage instances of JmlName.

6. As said earlier, top-level resolve method calls either
resolveNameRef or resolveFieldRef method. These two methods are quite
similar to each other in their structures. Both methods maintain
appropriate Reference objects as they scan expression characters. Only
the way of maintaining Reference objects are a bit different. As a
matter of fact, we could have merge these two methods considering
their similarity in exchange for reduced readability of code.

7. Special care needs to be taken for JML-specific expressions such as
a[0..2], a[*], and x.*. To handle a ranged array expression, we
extended ArrayReference with a range information to have
JmlRangeArrayReference. An instance of JmlRangeArrayReference keeps
low and high bounds of an array when an explicit range is provided as
in a[0..2]. If these bounds are not known as in a[*], this fact is
kept through an instance of AllRangeExpression.

Wild cards appended to an object are handled by the resolveWildCard
method. Unlike the case of the wild card in an array, it is fairly
straightforward to figure out all the members of an object. If x.* is
provided, all the members of x, including the members of its super
type, are inserted into the assignableReferences field.

8. One tricky issue worth noting is that in general a Reference
object's bit field should be reset before calling its resolveType
method. We found that otherwise an unnecessary error is reported. Method
reset does the job described above.

[JmlStoreRef.java]

1. moved this root class for store refs to under Reference because a
store ref denotes a set of locations in general as described in the
JML manual.

[JmlStoreRefInformalDescription.java, JmlStoreRefKeyword.java]

1. gave changes following the hierarchy change made in JmlStoreRef. As
mentioned before, its resolution will be dealt with in the next part
of the patch.

[JmlName.java]

1. Added the position field. This field is necessary when creating
instances of JmlFieldReference, SingleNameReference and
QualifiedNameReference in JmlStoreRefExpression. The value of the
position field is set based on the values of sourceStart and
sourceEnd.

2. Since the values of sourceStart and sourceEnd is, as described
above, required to set the value of field position, we changed the
signature of constructors of JmlName to include sourceStart and
sourceEnd.

3. Added methods to get array position and bounds.

[Scanner.java]

1. The original scanner failed to find token DOTDOT representing ".."
if ".." was preceded by a number mistaking this pattern as the cue for
floating number. In this process, one dot was unexpectedly lost. We
fixed this problem.

[Parser.java]

1. changed constructor calls to JmlStoreRefExpression and JmlName
following their constructor signature changes as described earlier.

[JmlRangeArrayReference.java]

1. added to handle JML-specific array expressions such as a[0..2] and
a[*].  

2. extended ArrayReference not JmlArrayReference because it is used
only in JML specifications.

[JmlRangeExpression.java]

1. added to handle ranged array expressions like a[0..2].

[JmlAllRangeExpression.java]

1. added to handle wild-card-embedded array expression like a[*].

====================================
<Test>

Some of existing test cases, mostly those in
AssignableTestCompiler.java conflict with the suggesting patch because
they are initially made to test only syntax of assignable clauses. We
made a proper change to dissolve the conflicts. We also detected one
test case in DbcTestCompiler.java was unfairly passed and made a
proper change as well. More details are in the following.

[DbcTestCompiler.java]

1. test_0032_implies_that: This test fails because \n in \nothing is
interpreted as a new line. A change to \\nothing dismisses the
problem.

[AssignableTestCompiler.java]

1. test_0002_AssignableWithIdentifier,
test_0003_AssignableWithIdentifierAndEnsures,
test_0004_AssignableWithIdentifiers,
test_0009_AssignableWithIdentifierSuffix,
test_0011_AssignableWithWildcardSuffix,
test_0012_AssignableWithArrayExpr, test_0013_AssignableWithArrayRange,
test_0014_AssignableWithArrayAll,
test_0015_AssignableWithMultipleSuffixes,
test_0016_AssignableWithArrayRange: the variables used in modifies
clause were not defined. We changed the tests to include proper
variable definitions.

2. test_0010_AssignableWithThisSuffix: Aside from the
undefined-variable issue described above, another issue around this
test case is the expression that ends with "this". Since now we are
considering semantics as well as syntax, we can say that this type of
expression should be avoided; it does not make sense to be able to
modify the receiver of a method. To avoid this expression, we changed
it to X.this.i. (An error reporting on this-ended expression will be
included in the next patch.)

August 15, 2008
-Jooyong Lee